Se actualiza detalle pedido
